libft

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: mel-bouh <mel-bouh@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/31 17:18:15 by mel-bouh          #+#    #+#             */
-/*   Updated: 2023/11/03 21:56:33 by mel-bouh         ###   ########.fr       */
+/*   Updated: 2023/11/06 23:30:25 by mel-bouh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
-	unsigned int	len_d;
-	unsigned int	len_s;
+	size_t	i;
+	size_t	len_d;
+	size_t	len_s;
 
 	i = 0;
 	len_d = ft_strlen(dst);
